+++ stages/deps_to_ninja/main.py
@@ -76,6 +76,7 @@ packages are not rebuilt unnecessarily.
 """
 
 from utilities import OutputDirectory, get_argparser
+from utilities import strip_version_info, interpret_bash_array
 
 from datetime import datetime
 from functools import partial
@@ -212,37 +213,6 @@ def build_triples(infos, args):
     return triples
 
 
-def interpret_bash_array(pkgbuild, array_name):
-    """Return the Bash array array_name in the file at path pkgbuild.
-
-    Returns:
-        A list of strings, or an empty string if the array is not
-        defined in the PKGBUILD, or None if there was a problem
-        interpreting the PKGBUILD.
-    """
-    cmd = dedent("""\
-                 #!/bin/bash
-                 . %s
-                 for foo in ${%s[@]}; do
-                    echo ${foo};
-                 done;
-                 """ % (pkgbuild, array_name))
-    with tempfile(mode="w") as temp:
-        temp.write(cmd)
-        temp.flush()
-        try:
-            cp = run(["/bin/bash", temp.name], stdout=PIPE,
-                     universal_newlines=True, timeout=20)
-        except TimeoutExpired:
-            print("Unable to interpret array %s in file %s" %
-                  (array_name, pkgbuild), file=stderr)
-            exit(1)
-
-    if not cp.stdout:
-        return []
-    else: return [line for line in cp.stdout.splitlines() if line]
-
-
 def gather_package_data(abs_dir,
         name_data, args, global_infos, provides):
 
@@ -270,20 +240,6 @@ def gather_package_data(abs_dir,
      for name in interpret_bash_array(pkgbuild, "provides")]
 
     global_infos.append(info)
-
-
-def strip_version_info(package_name):
-    """Return package_name without version numbers.
-
-    Some packages specified as dependencies have a version number, e.g.
-    gcc>=5.1. We shouldn't care about this, we always sync to an
-    up-to-date mirror before building packages, so strip this info.
-    """
-    for pat in [r">=", r"<=", r"=", r"<", r">"]:
-        depth = search(pat, package_name)
-        if depth:
-            package_name = package_name[:depth.start()]
-    return package_name
 
 
 def resolve_provides(package_infos, provides_dict):
